refactor(calendar): 서비스 레이어 분리 및 DI 체인 정리 - CalendarApi(GetConnect) / CalendarService / ScheduleService(어댑터) 구성 - MainBindings: CalendarApi→CalendarService→ScheduleService 순으로 등록 - 불필요 로그/주석 정리
